api-server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
mysql-connector-python==8.3.0
python-dotenv==1.0.0
docker==7.0.0
aiofiles==23.2.1
